+++ src/nespresso/db/models/tg_user.py
@@ -26,6 +26,8 @@ class TgUser(Base):
         Boolean, nullable=False, default=False
     )
 
+    about: Mapped[str] = mapped_column(String, nullable=True)
+
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True),
         server_default=text("CURRENT_TIMESTAMP"),