+++ CHANGELOG.md
@@ -26,6 +26,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [0.1.0] - 2025-06-11 - Vlad
+
+### Added
+
+- Matching Framework
+
 ## [0.1.0] - 2025-06-10 - Vlad
 
 ### Added